modules/connections-api/requirements.txt
mercator==0.1.11
shapely==1.7.0
SQLAlchemy==1.3.19
GeoAlchemy2==0.8.4
grpcio==1.44.0
grpcio-tools==1.44.0
psycopg2-binary==2.8.5
